cleaned up scripts and Dockerfiles

+++ srcs/requirements/wordpress/tools/wordpress_setup.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+if test -f "/var/www/html/wp-config.php"; then
+    exit 0
+fi
+
 mkdir -p /var/www/html
 cd /var/www/html
 rm -rf /var/www/html/*
@@ -10,6 +14,6 @@ mv wordpress/* .
 rm -rf latest.tar.gz
 rm -rf wordpress
 
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+/wordpress_config_setup.sh
 
 php -S 0.0.0.0:9000 -t .
